Set the last Alpha release as version
The version is not much relevant because the "set_version" helper
should be used to specify the binary package URL.
However, it still makes sense to have the latest SDK mentioned here
to spare the user a needless update. It also resolves any confusion
of the version.txt's relation to the development branches by using
the latest Alpha version available.