Rename flatcar-master-alpha to main and remove stale manifests
The development happens only on the new main branch and all channel
specific-branches are retired. The only exception will be Edge which
keeps its own development branch (but will be based on main).
The manifests for the channels' maintenance branches are now found as
maintenance.xml in the release tags/branches.