switch to flatcar-master branches and drop rkt
Some of our ebuilds specify the commit/revision to be built using
CROS_WORKON_COMMIT. The manifest specifies what branch of each third_party
repository to checkout. These two are completely separate. The clones are only
used as a reference, though `cros_workon` can be used to switch the ebuild into
a live (-9999) state and build against the checked-out sources.

The commits we specify with CROS_WORKON_COMMIT are on the flatcar-master
branch, so switch the manifest to that branch for most projects, so that the
correct sources are easily accessible.

Rkt has been removed from our images so we can drop that project.

Change-Id: Id4af44bc771392a2e5d1de6586fccb9343ec0a20
Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>